+++ src/emulator/core/Machine.h
@@ -3,6 +3,7 @@
 #include "emulator/core/Bus.h"
 #include "emulator/core/ICPU.h"
 #include "emulator/devices/CIA6526.h"
+#include "emulator/devices/VIC6566.h"
 #include "emulator/cpu/CPU8502.h"
 #include "emulator/cpu/CPU65C02.h"
 #include "emulator/devices/Memory.h"
@@ -49,8 +50,10 @@ public:
     Memory&  ram()  { return ram_; }
     CIA6526& cia1() { return cia1_; }
     CIA6526& cia2() { return cia2_; }
+    VIC6566& vic()  { return vic_; }
 
     const CIA6526& cia1() const { return cia1_; }
+    const VIC6566& vic()  const { return vic_;  }
 
     // -----------------------------------------------------------------------
     // Lifecycle
@@ -74,6 +77,7 @@ private:
     Memory  ram_;
     CIA6526 cia1_;
     CIA6526 cia2_;
+    VIC6566 vic_;
 
     CPU8502  cpu8502_;
     CPU65C02 cpu65c02_;